Adding template functions seperately as  plugins
