+++ pipewrench/merge.py
@@ -25,7 +25,7 @@ import os
 from json import dumps
 from jinja2 import Template
 import yaml
-
+from plugins import *
 
 
 OUT_DIR = "output"
@@ -34,7 +34,6 @@ LOGGING_FORMAT = "[%(levelname) s:%(filename)s:%(lineno)s in %(funcName)20s() ] 
 # Type mappings are loaded globally by `load_mapping_file` so they can be used from templates
 # pylint: disable=invalid-name
 type_mappings = {}
-
 
 def main():
     """
@@ -260,58 +259,6 @@ def write(string, fpath):
 if __name__ == '__main__':
     main()
 
-
-# Template functions.
-# These functions are intended to be called from Jinja2 templates.
-def map_datatypes(column):
-    """
-    Given a column, extract its datatype and return possible mappings
-     for it from a type-mappings.yml file.
-     For example, a mapping with the datatype 'bigint' may look like:
-
-    bigint:
-     kudu: bigint
-     impala: bigint
-     parquet: bigint
-     avro: long
-
-     when given a column with 'bigint' this function will return:
-
-     kudu: bigint
-     impala: bigint
-     parquet: bigint
-     avro: long
-
-    This function is intended to be called directly from templates
-    :param conf: The configuration. Not used but kept for consistency with other functions.
-    :param column: The column containing the datatype to map
-    :return: The mapped datatype
-    """
-    datatype = column['datatype'].lower()
-    logging.debug('found datatype %s', datatype)
-    mapped_datatype = type_mappings['type_mapping'].get(datatype)
-    logging.debug('mapped %s to %s', datatype, mapped_datatype)
-    return mapped_datatype
-
-
-def map_clobs(columns):
-    """
-    The map_clobs function will inject the following if there
-    are clob data types. --map-column-java col6=String,col8=String
-    This is needed as clob columns will not come into Hadoop unless
-    we specifically map them.
-    """
-    hasclobs = False
-    clobs = ""
-    for c in columns:
-        if c.get("datatype").lower() == "clob":
-            if not hasclobs:
-                hasclobs = True
-                clobs = "--map-column-java "
-            clobs = clobs + c.get("name") + "=String,"
-    return clobs[:-1]
-
-
 # Testing Functions
 def merge_single_template(template_file_path, type_mapping, conf):
     """
@@ -326,24 +273,3 @@ def merge_single_template(template_file_path, type_mapping, conf):
     with codecs.open(template_file_path, 'r', 'UTF-8') as template_file:
         template = template_file.read()
         return render(template, conf=conf, table=table)
-
-def order_columns(pks, columns):
-    """
-    Orders column list to include primary keys first and then non primary
-    key columns
-    :param pks: primary key list
-    :param columns: columns
-    :return: primary key columns + non primary key columns ordered
-    """
-    pk_list = []
-    non_pk_list = []
-
-    for c in columns:
-        for pk in pks:
-            if c.get("name") == pk:
-                pk_list.append(c)
-                break
-            elif pks[-1] == pk:
-                non_pk_list.append(c)
-
-    return pk_list+non_pk_list
