+++ pipewrench/plugins/order_columns.py
@@ -1,0 +1,34 @@
+#    Copyright 2017 Cargill Incorporated
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
+def order_columns(pks, columns):
+    """
+    Orders column list to include primary keys first and then non primary
+    key columns
+    :param pks: primary key list
+    :param columns: columns
+    :return: primary key columns + non primary key columns ordered
+    """
+    pk_list = []
+    non_pk_list = []
+
+    for c in columns:
+        for pk in pks:
+            if c.get("name") == pk:
+                pk_list.append(c)
+                break
+            elif pks[-1] == pk:
+                non_pk_list.append(c)
+
+    return pk_list+non_pk_list
